add Icons and create Edit Page

+++ src/views/Home.vue
@@ -2,11 +2,11 @@
   <div class="home">
     <ul>
       <li class="contact" v-for="contact in contacts">
-        <router-link :to="{ name: 'SingleContact', params: {id: contact.id, name: contact.name, surname: contact.surname, phone: contact.phone, email: contact.email, adress: contact.adress, birthday: contact.birthday} }">
+        <router-link class="contactlist" :to="{ name: 'SingleContact', params: {id: contact.id} }">
            <h3 class = "avatar">{{contact.name.charAt(0)}}{{contact.surname.charAt(0)}}</h3>
            <span class="info">{{contact.name}} {{contact.surname}}</span>
         </router-link>
-        <button class="deleteBtn" @click="deleteContact(contact)">X</button>
+        <button class="button" @click="deleteContact(contact)"><img class="icon" alt="trash-bin" src="../assets/trash-bin.png"/></button>
       </li>
     </ul>
   </div>
@@ -25,7 +25,10 @@ export default {
   },
   methods: {
     deleteContact(contact) {
-      this.$store.dispatch('deleteContact', contact);
+      let response = confirm(`Are you sure you want to delete contact ${contact.name}`)
+      if(response){
+        this.$store.dispatch('deleteContact', contact);
+      }
     }
   }
 }
@@ -47,7 +50,7 @@ export default {
     justify-content: center;
     align-items: center;
   }
-  .contact a{
+  .contactlist {
     text-decoration: none;
     display: flex;
     border: 2px solid transparent;
@@ -83,7 +86,7 @@ export default {
     margin-left: 20px;
   }
 
-  .deleteBtn {
+  .button {
     border: 1px solid transparent;
     background-color: #b22222;
     border-radius: 100%;
@@ -92,6 +95,14 @@ export default {
     height: 40px;
     margin-left: -30px;
     box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-
+    margin-right: 40px;
+  }
+  .icon {
+    height: 20px;
+    width: 20px;
+  }
+  .contactIcon {
+    height: 25px;
+    width: 25px;
   }
 </style>
